Scheduler: set initial values after scaling.
If the scaling operation did not succeed (and throws an exception), the
initial values will never be set, which is good.

Sidekiq will retry the job when an exception is thrown. Having the
initial values set before we actually scale (and where an exception
might occure), would result in `Slot#scaleable?` returning false,
meaning the scale opperation would never execute.