Slots: don't use enums for size.
Enums in Rails are hard to validate. An exception will be thrown when
you create a new object with the wrong value.

As we want to import data, we don't know what the user will give us as
data, so we need to be able to accept wrong data and validate that.